requirements.txt

python-telegram-bot==20.7
ccxt
pandas
numpy
ta
requests
python-dotenv
scikit-learn
